pyproject.toml
[tool.poetry]
name = "djangoproject1"
version = "0.1.0"
description = ""
authors = ["Zarnigor <zarnigorn1255@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
python-dotenv = "^1.0.1"
psycopg = "^3.2.3"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
